requirements.txt
pygame==1.9.6
mysql-connector==2.2.9